// src/utils/api/api.ts
export const ActivityLevel = {
    Weak: 'WEAK',
    Growing: 'GROWING',
    Strong: 'STRONG',
    Restricted: 'RESTRICTED'
} as const;

export type ActivityLevel = typeof ActivityLevel[keyof typeof ActivityLevel];
// ...
export interface ShipEngine {
    /**
     * The symbol of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'symbol': ShipEngineSymbolEnum;
    /**
     * The name of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'name': string;
    /**
     * The description of the engine.
     * @type {string}
     * @memberof ShipEngine
     */
    'description': string;
    /**
     * The repairable condition of a component. A value of 0 indicates the component needs significant repairs, while a value of 1 indicates the component is in near perfect condition. As the condition of a component is repaired, the overall integrity of the component decreases.
     * @type {number}
     * @memberof ShipEngine
     */
    'condition': number;
    /**
     * The overall integrity of the component, which determines the performance of the component. A value of 0 indicates that the component is almost completely degraded, while a value of 1 indicates that the component is in near perfect condition. The integrity of the component is non-repairable, and represents permanent wear over time.
     * @type {number}
     * @memberof ShipEngine
     */
    'integrity': number;
    /**
     * The speed stat of this engine. The higher the speed, the faster a ship can travel from one point to another. Reduces the time of arrival when navigating the ship.
     * @type {number}
     * @memberof ShipEngine
     */
    'speed': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipEngine
     */
    'requirements': ShipRequirements;
}

export const ShipEngineSymbolEnum = {
    ImpulseDriveI: 'ENGINE_IMPULSE_DRIVE_I',
    IonDriveI: 'ENGINE_ION_DRIVE_I',
    IonDriveIi: 'ENGINE_ION_DRIVE_II',
    HyperDriveI: 'ENGINE_HYPER_DRIVE_I'
} as const;

export type ShipEngineSymbolEnum = typeof ShipEngineSymbolEnum[keyof typeof ShipEngineSymbolEnum];

/**
 * The frame of the ship. The frame determines the number of modules and mounting points of the ship, as well as base fuel capacity. As the condition of the frame takes more wear, the ship will become more sluggish and less maneuverable.
 * @export
 * @interface ShipFrame
 */
export interface ShipFrame {
    /**
     * Symbol of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'symbol': ShipFrameSymbolEnum;
    /**
     * Name of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'name': string;
    /**
     * Description of the frame.
     * @type {string}
     * @memberof ShipFrame
     */
    'description': string;
    /**
     * The repairable condition of a component. A value of 0 indicates the component needs significant repairs, while a value of 1 indicates the component is in near perfect condition. As the condition of a component is repaired, the overall integrity of the component decreases.
     * @type {number}
     * @memberof ShipFrame
     */
    'condition': number;
    /**
     * The overall integrity of the component, which determines the performance of the component. A value of 0 indicates that the component is almost completely degraded, while a value of 1 indicates that the component is in near perfect condition. The integrity of the component is non-repairable, and represents permanent wear over time.
     * @type {number}
     * @memberof ShipFrame
     */
    'integrity': number;
    /**
     * The amount of slots that can be dedicated to modules installed in the ship. Each installed module take up a number of slots, and once there are no more slots, no new modules can be installed.
     * @type {number}
     * @memberof ShipFrame
     */
    'moduleSlots': number;
    /**
     * The amount of slots that can be dedicated to mounts installed in the ship. Each installed mount takes up a number of points, and once there are no more points remaining, no new mounts can be installed.
     * @type {number}
     * @memberof ShipFrame
     */
    'mountingPoints': number;
    /**
     * The maximum amount of fuel that can be stored in this ship. When refueling, the ship will be refueled to this amount.
     * @type {number}
     * @memberof ShipFrame
     */
    'fuelCapacity': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipFrame
     */
    'requirements': ShipRequirements;
}

export const ShipFrameSymbolEnum = {
    Probe: 'FRAME_PROBE',
    Drone: 'FRAME_DRONE',
    Interceptor: 'FRAME_INTERCEPTOR',
    Racer: 'FRAME_RACER',
    Fighter: 'FRAME_FIGHTER',
    Frigate: 'FRAME_FRIGATE',
    Shuttle: 'FRAME_SHUTTLE',
    Explorer: 'FRAME_EXPLORER',
    Miner: 'FRAME_MINER',
    LightFreighter: 'FRAME_LIGHT_FREIGHTER',
    HeavyFreighter: 'FRAME_HEAVY_FREIGHTER',
    Transport: 'FRAME_TRANSPORT',
    Destroyer: 'FRAME_DESTROYER',
    Cruiser: 'FRAME_CRUISER',
    Carrier: 'FRAME_CARRIER'
} as const;

export type ShipFrameSymbolEnum = typeof ShipFrameSymbolEnum[keyof typeof ShipFrameSymbolEnum];
// ...
export interface ShipModule {
    /**
     * The symbol of the module.
     * @type {string}
     * @memberof ShipModule
     */
    'symbol': ShipModuleSymbolEnum;
    /**
     * Modules that provide capacity, such as cargo hold or crew quarters will show this value to denote how much of a bonus the module grants.
     * @type {number}
     * @memberof ShipModule
     */
    'capacity'?: number;
    /**
     * Modules that have a range will such as a sensor array show this value to denote how far can the module reach with its capabilities.
     * @type {number}
     * @memberof ShipModule
     */
    'range'?: number;
    /**
     * Name of this module.
     * @type {string}
     * @memberof ShipModule
     */
    'name': string;
    /**
     * Description of this module.
     * @type {string}
     * @memberof ShipModule
     */
    'description': string;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipModule
     */
    'requirements': ShipRequirements;
}
// ...
export type ShipModuleSymbolEnum = typeof ShipModuleSymbolEnum[keyof typeof ShipModuleSymbolEnum];

/**
 * A mount is installed on the exterier of a ship.
 * @export
 * @interface ShipMount
 */
export interface ShipMount {
    /**
     * Symbo of this mount.
     * @type {string}
     * @memberof ShipMount
     */
    'symbol': ShipMountSymbolEnum;
    /**
     * Name of this mount.
     * @type {string}
     * @memberof ShipMount
     */
    'name': string;
    /**
     * Description of this mount.
     * @type {string}
     * @memberof ShipMount
     */
    'description'?: string;
    /**
     * Mounts that have this value, such as mining lasers, denote how powerful this mount\'s capabilities are.
     * @type {number}
     * @memberof ShipMount
     */
    'strength'?: number;
    /**
     * Mounts that have this value denote what goods can be produced from using the mount.
     * @type {Array<string>}
     * @memberof ShipMount
     */
    'deposits'?: Array<ShipMountDepositsEnum>;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipMount
     */
    'requirements': ShipRequirements;
}
// ...
export type ShipMountSymbolEnum = typeof ShipMountSymbolEnum[keyof typeof ShipMountSymbolEnum];
// ...
export type ShipMountDepositsEnum = typeof ShipMountDepositsEnum[keyof typeof ShipMountDepositsEnum];
// ...
export interface ShipReactor {
    /**
     * Symbol of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'symbol': ShipReactorSymbolEnum;
    /**
     * Name of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'name': string;
    /**
     * Description of the reactor.
     * @type {string}
     * @memberof ShipReactor
     */
    'description': string;
    /**
     * The repairable condition of a component. A value of 0 indicates the component needs significant repairs, while a value of 1 indicates the component is in near perfect condition. As the condition of a component is repaired, the overall integrity of the component decreases.
     * @type {number}
     * @memberof ShipReactor
     */
    'condition': number;
    /**
     * The overall integrity of the component, which determines the performance of the component. A value of 0 indicates that the component is almost completely degraded, while a value of 1 indicates that the component is in near perfect condition. The integrity of the component is non-repairable, and represents permanent wear over time.
     * @type {number}
     * @memberof ShipReactor
     */
    'integrity': number;
    /**
     * The amount of power provided by this reactor. The more power a reactor provides to the ship, the lower the cooldown it gets when using a module or mount that taxes the ship\'s power.
     * @type {number}
     * @memberof ShipReactor
     */
    'powerOutput': number;
    /**
     * 
     * @type {ShipRequirements}
     * @memberof ShipReactor
     */
    'requirements': ShipRequirements;
}

export const ShipReactorSymbolEnum = {
    SolarI: 'REACTOR_SOLAR_I',
    FusionI: 'REACTOR_FUSION_I',
    FissionI: 'REACTOR_FISSION_I',
    ChemicalI: 'REACTOR_CHEMICAL_I',
    AntimatterI: 'REACTOR_ANTIMATTER_I'
} as const;

export type ShipReactorSymbolEnum = typeof ShipReactorSymbolEnum[keyof typeof ShipReactorSymbolEnum];
// ...
export interface ShipRequirements {
    /**
     * The amount of power required from the reactor.
     * @type {number}
     * @memberof ShipRequirements
     */
    'power'?: number;
    /**
     * The number of crew required for operation.
     * @type {number}
     * @memberof ShipRequirements
     */
    'crew'?: number;
    /**
     * The number of module slots required for installation.
     * @type {number}
     * @memberof ShipRequirements
     */
    'slots'?: number;
}
// ...
export const ShipType = {
    Probe: 'SHIP_PROBE',
    MiningDrone: 'SHIP_MINING_DRONE',
    SiphonDrone: 'SHIP_SIPHON_DRONE',
    Interceptor: 'SHIP_INTERCEPTOR',
    LightHauler: 'SHIP_LIGHT_HAULER',
    CommandFrigate: 'SHIP_COMMAND_FRIGATE',
    Explorer: 'SHIP_EXPLORER',
    HeavyFreighter: 'SHIP_HEAVY_FREIGHTER',
    LightShuttle: 'SHIP_LIGHT_SHUTTLE',
    OreHound: 'SHIP_ORE_HOUND',
    RefiningFreighter: 'SHIP_REFINING_FREIGHTER',
    Surveyor: 'SHIP_SURVEYOR'
} as const;

export type ShipType = typeof ShipType[keyof typeof ShipType];
// ...
export interface ShipyardShip {
    /**
     * 
     * @type {ShipType}
     * @memberof ShipyardShip
     */
    'type': ShipType;
    /**
     * 
     * @type {string}
     * @memberof ShipyardShip
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof ShipyardShip
     */
    'description': string;
    /**
     * 
     * @type {SupplyLevel}
     * @memberof ShipyardShip
     */
    'supply': SupplyLevel;
    /**
     * 
     * @type {ActivityLevel}
     * @memberof ShipyardShip
     */
    'activity'?: ActivityLevel;
    /**
     * 
     * @type {number}
     * @memberof ShipyardShip
     */
    'purchasePrice': number;
    /**
     * 
     * @type {ShipFrame}
     * @memberof ShipyardShip
     */
    'frame': ShipFrame;
    /**
     * 
     * @type {ShipReactor}
     * @memberof ShipyardShip
     */
    'reactor': ShipReactor;
    /**
     * 
     * @type {ShipEngine}
     * @memberof ShipyardShip
     */
    'engine': ShipEngine;
    /**
     * 
     * @type {Array<ShipModule>}
     * @memberof ShipyardShip
     */
    'modules': Array<ShipModule>;
    /**
     * 
     * @type {Array<ShipMount>}
     * @memberof ShipyardShip
     */
    'mounts': Array<ShipMount>;
    /**
     * 
     * @type {ShipyardShipCrew}
     * @memberof ShipyardShip
     */
    'crew': ShipyardShipCrew;
}


/**
 * 
 * @export
 * @interface ShipyardShipCrew
 */
export interface ShipyardShipCrew {
    /**
     * 
     * @type {number}
     * @memberof ShipyardShipCrew
     */
    'required': number;
    /**
     * 
     * @type {number}
     * @memberof ShipyardShipCrew
     */
    'capacity': number;
}
// ...
export const SupplyLevel = {
    Scarce: 'SCARCE',
    Limited: 'LIMITED',
    Moderate: 'MODERATE',
    High: 'HIGH',
    Abundant: 'ABUNDANT'
} as const;

export type SupplyLevel = typeof SupplyLevel[keyof typeof SupplyLevel];